fix: refactor sample
